Temporal creation of kcoin package, expecting to move it to kcoin folder.
