+++ cmd/genesis/kcoin/testfiles/testnet_default.json
